Refactored `EcdsaSignature` into `composable_support::types`
