+++ frame/crowdloan-rewards/src/models.rs
@@ -1,5 +1,5 @@
 use codec::{Decode, Encode, MaxEncodedLen};
-use composable_support::types::EthereumAddress;
+use composable_support::types::{EcdsaSignature, EthereumAddress};
 use scale_info::TypeInfo;
 use sp_runtime::{MultiSignature, RuntimeDebug};
 
@@ -21,19 +21,4 @@ pub enum Proof<AccountId> {
 pub enum RemoteAccount<AccountId> {
 	RelayChain(AccountId),
 	Ethereum(EthereumAddress),
-}
-
-#[derive(Encode, Decode, Clone, MaxEncodedLen, TypeInfo)]
-pub struct EcdsaSignature(pub [u8; 65]);
-
-impl PartialEq for EcdsaSignature {
-	fn eq(&self, other: &Self) -> bool {
-		self.0[..] == other.0[..]
-	}
-}
-
-impl sp_std::fmt::Debug for EcdsaSignature {
-	fn fmt(&self, f: &mut sp_std::fmt::Formatter<'_>) -> sp_std::fmt::Result {
-		write!(f, "EcdsaSignature({:?})", &self.0[..])
-	}
 }
